fix(realtime): keep sub-second transcript offsets, unhang the audio suite
Audio chunks are 2.5 seconds apart, so every offset ends in .5 and an INT
column silently rounded subtitles half a second away from the audio. Store
start_sec as DOUBLE, which mysql2 also returns as a number (DECIMAL comes
back as a string).

Two problems in the test itself: it released the first chunk before the
server had received the second, so it measured the case it meant to exclude;
and it left WebSockets open, which kept server.close() waiting forever.

Full backend suite now passes: 77 tests, 10 suites.

+++ backend/src/database/schema.sql
@@ -75,7 +75,9 @@ CREATE TABLE IF NOT EXISTS transcript_segments (
   meeting_id   VARCHAR(64)  NOT NULL,
   speaker_id   VARCHAR(64)  NOT NULL,
   speaker_name VARCHAR(255) NOT NULL,
-  start_sec    INT          NOT NULL,
+  -- ก้อนเสียงห่างกัน 2.5 วินาที offset จึงลงท้ายด้วย .5 เสมอ INT ตัดทศนิยมทิ้งแล้วคำบรรยาย
+  -- คลาดจากเสียงจริงได้ถึงครึ่งวินาที DOUBLE คืนค่าเป็น number ให้ mysql2 ต่างจาก DECIMAL ที่คืนสตริง
+  start_sec    DOUBLE       NOT NULL,
   text         TEXT         NOT NULL,
   created_at   BIGINT       NOT NULL,
   INDEX idx_transcript_meeting (meeting_id, start_sec)
